requirements.txt
bs4
feedparser
python-telegram-bot
boto3